features - 添加了 liker 表 - 添加了邮件模块，新增发邮件功能 - 添加了点赞功能、帖子更新提醒功能(notifier)、热度排序功能(top-k problem)

+++ handlers/user/upload_resume.go
@@ -31,7 +31,7 @@ func UploadResumeHandler(c *gin.Context) {
 	req := &UploadFileRequest{
 		File:     resume,
 		FileName: header.Filename,
-		UserID:   int64(userID),
+		UserID:   userID,
 	}
 
 	// 获取上传文件的响应
